+++ packages/zowe-explorer-api/__tests__/__unit__/profiles/ZoweExplorerZosmfApi.unit.test.ts
@@ -751,6 +751,21 @@ describe("ZosmfMvsApi", () => {
             expect(result).toStrictEqual({ count: 4, lastItem: "DS4" });
         });
     });
+
+    it("should test that resolve alias calls the zosmf API", async () => {
+        const dsName = "ALIAS.DS";
+        const resolveAlias = vi.fn((_session, ds, _options) => {
+            expect(ds).toEqual(dsName);
+            return { apiResponse: { targetDsn: "WOW.THE.ORIGINAL" }, commandResponse: "", success: true };
+        });
+
+        vi.spyOn(zosfiles.List, "resolveAlias").mockImplementation(resolveAlias as any);
+
+        const api = new ZoweExplorerZosmf.MvsApi();
+        api.getSession = vi.fn();
+        const resolved = await api.resolveAlias(dsName);
+        expect(resolved.apiResponse.targetDsn).toEqual("WOW.THE.ORIGINAL");
+    });
 });
 
 describe("ZosmfJesApi", () => {