+++ packages/zowe-explorer/src/trees/dataset/ZoweDatasetNode.ts
@@ -73,6 +73,8 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
     public favoritedMemberNames?: string[];
     public wasPds?: boolean;
     public justRecalled?: boolean;
+    public isAlias?: boolean;
+    public aliasTargetDsn?: string;
 
     private paginator?: Paginator<IZosFilesResponse>;
     private paginatorData?: {
@@ -99,6 +101,10 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
         } else {
             this.contextValue = isBinary ? Constants.DS_DS_BINARY_CONTEXT : Constants.DS_DS_CONTEXT;
         }
+        if (opts.isAlias) {
+            this.contextValue += Constants.DS_ALIAS_SUFFIX;
+            this.isAlias = true;
+        }
         if (opts.contextOverride?.includes(Constants.DS_SESSION_CONTEXT)) {
             const toolTipList: string[] = [];
             toolTipList.push(`${vscode.l10n.t("Profile: ")}${opts.label}`);
@@ -106,6 +112,10 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
             this.tooltip = toolTipList.join("\n");
         } else {
             this.tooltip = this.label as string;
+        }
+        if (opts.aliasTargetDsn) {
+            this.tooltip += `\n${vscode.l10n.t("Alias to: ")}${opts.aliasTargetDsn}`;
+            this.aliasTargetDsn = opts.aliasTargetDsn;
         }
         const icon = IconGenerator.getIconByNode(this);
         if (icon) {
@@ -132,17 +142,17 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
         if (this.label !== vscode.l10n.t("Favorites") && !SharedContext.isMigrated(this)) {
             const sessionLabel = opts.profile?.name ?? SharedUtils.getSessionLabel(this);
             if (
-                this.contextValue === Constants.DS_DS_CONTEXT ||
-                this.contextValue === Constants.DS_FAV_CONTEXT ||
-                this.contextValue === Constants.DS_PDS_CONTEXT ||
-                this.contextValue === Constants.PDS_FAV_CONTEXT ||
+                this.contextValue.startsWith(Constants.DS_DS_CONTEXT) ||
+                this.contextValue.startsWith(Constants.DS_FAV_CONTEXT) ||
+                this.contextValue.startsWith(Constants.DS_PDS_CONTEXT) ||
+                this.contextValue.startsWith(Constants.PDS_FAV_CONTEXT) ||
                 this.contextValue === Constants.VSAM_CONTEXT
             ) {
                 this.resourceUri = vscode.Uri.from({
                     scheme: ZoweScheme.DS,
                     path: `/${sessionLabel}/${this.label as string}`,
                 });
-                if (this.contextValue === Constants.DS_DS_CONTEXT || this.contextValue === Constants.DS_FAV_CONTEXT) {
+                if (this.contextValue.startsWith(Constants.DS_DS_CONTEXT) || this.contextValue.startsWith(Constants.DS_FAV_CONTEXT)) {
                     const extension = DatasetUtils.getExtension(this.label as string);
                     this.resourceUri = this.resourceUri.with({ path: `${this.resourceUri.path}${extension ?? ""}` });
                     this.command = { command: "vscode.open", title: "", arguments: [this.resourceUri] };
@@ -226,7 +236,7 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
      * Updates this node so the recalled data set can be interacted with.
      * @param isPds Whether the data set is a PDS
      */
-    public async datasetRecalled(isPds: boolean): Promise<void> {
+    public datasetRecalled(isPds: boolean): void {
         // Change context value to match dsorg, update collapsible state and assign resource URI
         // Preserve favorite context and any additional context values
         this.contextValue = this.contextValue.replace(Constants.DS_MIGRATED_FILE_CONTEXT, isPds ? Constants.DS_PDS_CONTEXT : Constants.DS_DS_CONTEXT);
@@ -250,7 +260,6 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
             this.command = { command: "vscode.open", title: "", arguments: [this.resourceUri] };
             if (!DatasetFSProvider.instance.exists(this.resourceUri)) {
                 DatasetFSProvider.instance.createEntry(this.resourceUri, DsType.Ps);
-                await vscode.workspace.fs.writeFile(this.resourceUri, new Uint8Array());
             }
         }
     }
@@ -291,20 +300,20 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
      * @param dsorg The dataset organization (e.g., "PO" for PDS)
      * @param dsTree The dataset tree provider
      */
-    private async syncNodeMigrationStatus(dsNode: ZoweDatasetNode, migr: string, dsorg: string, dsTree: DatasetTree): Promise<void> {
+    private syncNodeMigrationStatus(dsNode: ZoweDatasetNode, migr: string, dsorg: string, dsTree: DatasetTree): void {
         const migrationStatus = migr.toUpperCase();
         if (migrationStatus !== "YES") {
             dsNode.justRecalled = false;
         }
         if (SharedContext.isMigrated(dsNode) && migrationStatus !== "YES") {
             const isPds = dsorg?.startsWith("PO") ?? dsNode.wasPds ?? false;
-            await dsNode.datasetRecalled(isPds);
+            dsNode.datasetRecalled(isPds);
             if (dsTree) {
                 const isFav = SharedContext.isFavoriteDescendant(dsNode);
                 const equiv = dsTree.findEquivalentNode(dsNode, isFav) as ZoweDatasetNode;
                 if (equiv) {
                     equiv.justRecalled = false;
-                    await equiv.datasetRecalled(isPds);
+                    equiv.datasetRecalled(isPds);
                     dsTree.refreshElement(equiv.getParent() as IZoweDatasetTreeNode);
                 }
                 if (isFav || (equiv && SharedContext.isFavoriteDescendant(equiv))) {
@@ -402,7 +411,7 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
                 if (dsNode != null) {
                     elementChildren[dsNode.label.toString()] = dsNode;
                     if (item.migr) {
-                        await this.syncNodeMigrationStatus(dsNode, item.migr, item.dsorg, dsTree);
+                        this.syncNodeMigrationStatus(dsNode, item.migr, item.dsorg, dsTree);
                     }
                 } else if (item.migr && item.migr.toUpperCase() === "YES") {
                     // Creates a ZoweDatasetNode for a migrated dataset
@@ -414,6 +423,26 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
                         profile: cachedProfile,
                     });
                     dsNode.wasPds = item.dsorg?.startsWith("PO");
+                    elementChildren[dsNode.label.toString()] = dsNode;
+                } else if (item.vol === "*ALIAS") {
+                    const resolvedAlias = await this.resolveAlias(item);
+                    item.dsorg ??= resolvedAlias?.dsorg;
+                    item.recfm ??= resolvedAlias?.recfm;
+                    item.blksz ??= resolvedAlias?.blksz;
+                    item.migr ??= resolvedAlias?.migr;
+                    const originalIsMigrated = item.migr?.toUpperCase() === "YES";
+                    dsNode = new ZoweDatasetNode({
+                        label: item.dsname,
+                        collapsibleState:
+                            originalIsMigrated || !item.dsorg?.startsWith("PO")
+                                ? vscode.TreeItemCollapsibleState.None
+                                : vscode.TreeItemCollapsibleState.Collapsed,
+                        parentNode: this,
+                        profile: cachedProfile,
+                        isAlias: true,
+                        aliasTargetDsn: resolvedAlias?.dsname,
+                        contextOverride: originalIsMigrated ? Constants.DS_MIGRATED_FILE_CONTEXT : undefined,
+                    });
                     elementChildren[dsNode.label.toString()] = dsNode;
                 } else if (item.dsorg?.startsWith("PO")) {
                     // Creates a ZoweDatasetNode for a PDS
@@ -880,6 +909,42 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
         };
     }
 
+    private async resolveAlias(item: IZosmfListResponse): Promise<IZosmfListResponse | undefined> {
+        const profile = Profiles.getInstance()?.loadNamedProfile(this.getProfile().name);
+        const mvsApi = ZoweExplorerApiRegister.getMvsApi(profile);
+        if (mvsApi.resolveAlias) {
+            try {
+                const resolution = await mvsApi.resolveAlias(item.dsname);
+                if (resolution.apiResponse.targetDsn) {
+                    const originalAttributes = await mvsApi.dataSet(resolution.apiResponse.targetDsn, { attributes: true });
+                    ZoweLogger.info(
+                        `[ZoweDatasetNode.resolveAlias] Resolved alias ${item.dsname} to ${resolution.apiResponse.targetDsn}.` +
+                            ` Retrieving attributes of the original data set.`
+                    );
+                    const matchingOriginal = (originalAttributes.apiResponse.items ?? originalAttributes.apiResponse) as IZosmfListResponse[];
+                    if (matchingOriginal.length > 0) {
+                        return {
+                            ...item,
+                            dsorg: matchingOriginal[0].dsorg,
+                            recfm: matchingOriginal[0].recfm,
+                            blksz: matchingOriginal[0].blksz,
+                            dsname: resolution.apiResponse.targetDsn,
+                            migr: matchingOriginal[0].migr,
+                        };
+                    }
+                }
+            } catch (e) {
+                ZoweLogger.error("[ZoweDatasetNode.resolveAlias] Encountered an error trying to resolve the alias: " + e.message);
+            }
+        } else {
+            ZoweLogger.warn(
+                `[ZoweDatasetNode.resolveAlias] MVS API for ${profile.type} does not implement resolveAlias.` +
+                    ` Alias ${item.dsname} will not be resolved.`
+            );
+        }
+        return undefined;
+    }
+
     public async listDatasets(responses: IZosFilesResponse[], options?: Definitions.DatasetListOpts): Promise<void> {
         const dsPatterns = [
             ...new Set(
@@ -1015,11 +1080,11 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
             this.memberPattern = this.memberPattern.toUpperCase();
             for (const memPattern of this.memberPattern.split(",")) {
                 baseOptions.pattern = memPattern.trim();
-                responses.push(await mvsApi.allMembers(this.label as string, baseOptions));
+                responses.push(await mvsApi.allMembers(this.aliasTargetDsn ?? (this.label as string), baseOptions));
             }
         } else {
             // Fetching members for this PDS
-            responses.push(await mvsApi.allMembers(this.label as string, baseOptions));
+            responses.push(await mvsApi.allMembers(this.aliasTargetDsn ?? (this.label as string), baseOptions));
         }
     }